refactor(web): extract useOverviewData hook from Overview.tsx — T10
Move all derived-state computation (useMemo, useEffect, useState) into
a dedicated useOverviewData hook. Overview.tsx drops from 509 to 140 LOC.
No behaviour change — JSX and prop-passing are identical.
